refactor(Settings): Put settings related files in src/Settings
These files are moved to src/Settings

1. src/Preferences/*
2. src/Core/SettingsManager.*
3. setting.json (moved to src/Settings/settings.json)
