--- src/Preferences/README.md
@@ -1,3 +1,0 @@
-# Extensions
-
-This folder contains the codes for the preferences window.
