+++ src/Settings/PathItem.hpp
