+++ src/Settings/README.md
@@ -1,0 +1,3 @@
+# Extensions
+
+This folder contains the codes related to the settings.
