+++ src/Settings/SettingsManager.hpp
