+++ src/Settings/ShortcutItem.hpp
